ml-pipeline/requirements.txt
pandas
numpy
scikit-learn
xgboost
optuna
kagglehub
psycopg2-binary
matplotlib
